SPA (#4)
* refactor: project converted to a next.js project with python

* add: husky

* fix: page styling and dark mode debugged

* style: format

* add: dark mode and red theming applying

Some tailwind code forcing colors has been changed to generic color variables

* add: spa skeleton defined

* add: prisma

* style: printWidth=120

* add: navbar updated

* fix: Prompt input and submit button style adjusted

* add: onSubmit implemented

* add: question grid

Server keeps colliding, it may be caused by the usage of two apis at the same time. I will try to fix it as soon as possible migrating langchain code to js instead of python.

* fix: simplified the webpage to entirely work and then build up

* add: last update info

* fix: db update

* fix: .gitkeep kept

+++ f1guru/database.py
@@ -1,5 +1,6 @@
 """This module provides functions to interact with the F1DB database."""
 
+import subprocess
 from typing import Any, Dict, Sequence
 import pymysql
 import requests
@@ -15,30 +16,16 @@ load_dotenv()
 
 def get_db() -> SQLDatabase:
     """Get a SQLAlchemy database wrapper for the F1DB database."""
-    user = os.environ.get("TIDB_USER")
-    password = os.environ.get("TIDB_PASSWORD")
-    host = os.environ.get("TIDB_HOST")
-    port = os.environ.get("TIDB_PORT")
-    db_name = os.environ.get("TIDB_DB_NAME")
+    user = os.environ.get("F1DB_USER")
+    password = os.environ.get("F1DB_PASSWORD")
+    host = os.environ.get("F1DB_HOST")
+    port = os.environ.get("F1DB_PORT")
+    db_name = os.environ.get("F1DB_DB_NAME")
 
     tidb_connection_string = f"mysql+pymysql://{user}:{password}@{host}:{port}/{db_name}?ssl_ca=/etc/ssl/cert.pem&ssl_verify_cert=true&ssl_verify_identity=true"
     db = SQLDatabase.from_uri(tidb_connection_string)
 
     return db
-
-
-def get_db_connection() -> pymysql.connections.Connection:
-    """Get a pymysql database connection to the F1DB database."""
-    return pymysql.connect(
-        host=os.environ.get("TIDB_HOST"),
-        port=os.environ.get("TIDB_PORT"),
-        user=os.environ.get("TIDB_USER"),
-        password=os.environ.get("TIDB_PASSWORD"),
-        database=os.environ.get("TIDB_DB_NAME"),
-        ssl_ca="/etc/ssl/cert.pem",
-        ssl_verify_cert=True,
-        ssl_verify_identity=True,
-    )
 
 
 def run_query(query: str) -> str | Sequence[Dict[str, Any]] | Result[Any]:
@@ -67,6 +54,9 @@ def check_for_new_release() -> None:
     if response.status_code == 200:
         latest_release = response.json()
         if os.environ.get("LAST_F1DB_RELEASE") != latest_release["tag_name"]:
+            print(
+                f"New release found. Rebuilding database from release {os.environ.get('LAST_F1DB_RELEASE')} to {latest_release['tag_name']}..."
+            )
             rebuild_db_from_release(latest_release)
             os.environ["LAST_F1DB_RELEASE"] = latest_release["tag_name"]
 
@@ -86,7 +76,8 @@ def rebuild_db_from_release(release_info: dict) -> None:
         update_database(join(dirname(__file__), "temp", "f1db-sql-mysql.sql"))
 
         for file in os.listdir(join(dirname(__file__), "temp")):
-            os.remove(join(dirname(__file__), "temp", file))
+            if file != ".gitkeep":
+                os.remove(join(dirname(__file__), "temp", file))
     elif len(asset) > 1:
         raise ValueError("Multiple assets found with the same name")
     else:
@@ -106,17 +97,7 @@ def download_file(url: str) -> str:
 
 def update_database(sql_file_path: str) -> None:
     """Update the F1DB database with the SQL file at the given path."""
-    try:
-        connection = get_db_connection()
-        cursor = connection.cursor()
+    # mysql -h {os.environ.get("F1DB_HOST")} -P {os.environ.get("F1DB_PORT")} -u {os.environ.get("F1DB_USER")} -p{os.environ.get("F1DB_PASSWORD")} {os.environ.get("F1DB_DB_NAME")} < {sql_file_path}
+    command = f"""mysql -h {os.environ.get("F1DB_HOST")} -P {os.environ.get("F1DB_PORT")} -u {os.environ.get("F1DB_USER")} -p{os.environ.get('F1DB_PASSWORD')} {os.environ.get("F1DB_DB_NAME")} < {sql_file_path}"""
 
-        with open(sql_file_path, "r") as f:
-            sql_script = f.read()
-
-        cursor.execute(sql_script)
-        connection.commit()
-    except pymysql.Error as e:
-        raise e
-    finally:
-        if connection:
-            connection.close()
+    subprocess.run(command, shell=True)
